Completed User Operations
Sign up button allows for a new user to be created given email, username and password. It will then sign the user in automatically.

Login will sign in the user given username and password.

Sign out will remove the cookies from the users client and will reload verifying that the user is no longer connected to an account.

Users page displays a list of all users using a table with a key to the users ID.

User can change either their username or their password and will need to refresh to see the username change, or for the password change the user would need to relog to see confirm effects.

User can also delete account in the same profile page and will automatically send them back to the sign in page.

Cart table shows the cart object for the person logged on.